+++ scrm-service/src/main/java/com/huotu/scrm/service/entity/report/DayReport.java
@@ -45,7 +45,7 @@ public class DayReport {
      * 等级
      */
     @Column(name = "Level_Id")
-    private int levelId;
+    private Long levelId;
 
     /**
      * 是否是销售员（1：是 0：不是）
